a script that deletes all State objects with a name containing the letter a from the database

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -16,8 +16,8 @@ if __name__ == "__main__":
     Base.metadata.create_all(eng)
     Session = sessionmaker(bind=eng)
     session = Session()
-    states = session.query(State).filter(State.name.like('%a%'))
-    for state in states:
-        session.delete(state)
+    for state in session.query(State):
+        if "a" in state.name:
+            session.delete(state)
     session.commit()
     session.close()
